Add option to make Return key insert LF or CR LF instead of CR
Many serial devices take text commands terminated with LF or CR LF, but
by default PuTTY maps the Return key to CR. The Return key can be
remapped to CR LF if the device sends the escape sequence "\e[20h", but
this requires the device to have something like an "echo" command, and
even if it does, there is still no escape sequence to map the Return key
to just LF.

This patch adds a setting to control the Return key's behavior when the
terminal starts. After that, the escape sequence "\e[20l" will change
the setting to CR and "\e[20h" will change it to CR LF. If the terminal
is started in LF mode and then an escape sequence changes the mode,
there is no escape sequence to get back to LF mode, but if one of those
escape sequences was sent then LF is the wrong mode anyway.